set errorbar to 1 pixel
